+++ packages/lzards-api-client/src/index.ts
@@ -1,4 +1,4 @@
 export {
-  sendGetRequestToLzards,
+  submitQueryToLzards,
   getAuthToken,
 } from './lzards';